python-admin/requirements.txt
# Python Admin Desktop App Requirements
# Core dependencies for the blog post management application

# GUI Framework
pillow>=10.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Database and API
supabase>=2.0.0

# HTTP Requests
requests>=2.31.0

# AWS S3 SDK for Cloudflare R2
boto3>=1.34.0

# Date and Time handling
python-dateutil>=2.8.0

# Text processing and validation
python-slugify>=8.0.0
markdown>=3.5.0

# Security and encryption
cryptography>=41.0.0
